# restapi/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A set of APIs to allow applications interact to with the Umay API.",
        "title": "Umay API",
        "contact": {
            "name": "Maydere",
            "url": "http://www.maydere.com",
            "email": "destek@maydere.com"
        },
        "version": "1.0"
    },
    "host": "localhost:1919",
    "basePath": "/v1",
    "paths": {
        "/common/countries": {
            "get": {
                "description": "Get countries.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Get countries",
                "operationId": "get-countries",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.CountriesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/common/countries/{id}": {
            "get": {
                "description": "Get country.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Get country",
                "operationId": "get-country",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.CountryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/common/languages": {
            "get": {
                "description": "Get languages.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Get languages",
                "operationId": "get-languages",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.LanguagesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/common/languages/{id}": {
            "get": {
                "description": "Get language.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Common"
                ],
                "summary": "Get language",
                "operationId": "get-language",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.LanguageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "pb.CountriesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Country"
                    }
                }
            }
        },
        "pb.Country": {
            "type": "object",
            "properties": {
                "calling_code": {
                    "type": "integer"
                },
                "display_order": {
                    "type": "integer"
                },
                "english_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "iso_code2": {
                    "type": "string"
                },
                "iso_code3": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "published": {
                    "type": "boolean"
                }
            }
        },
        "pb.CountryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/pb.Country"
                }
            }
        },
        "pb.Language": {
            "type": "object",
            "properties": {
                "culture_code": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "published": {
                    "type": "boolean"
                },
                "rtl": {
                    "type": "boolean"
                }
            }
        },
        "pb.LanguageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/pb.Language"
                }
            }
        },
        "pb.LanguagesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Language"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}